engine/scripts/requirements.txt
lightgbm>=4.3.0
scikit-learn>=1.5.0
pandas>=2.2.0
torch>=2.3.0
sentence-transformers>=2.7.0
joblib>=1.4.0
pikepdf>=9.0.0
pdfminer.six>=20231228
lxml>=5.2.0
python-magic-bin; platform_system=="Windows"
python-magic; platform_system!="Windows"
